+++ source/Age/Internal/BvhTree.cs
@@ -199,7 +199,7 @@ public unsafe class BvhTree
             {
                 Border         = new Border(2, 0, color * new Color(1, 1, 1, 1)),
                 Flags          = Shaders.CanvasShader.Flags.ColorAsBackground,
-                Size           = new(bvhNode.AABB.Size.X, bvhNode.AABB.Size.Y),
+                Size           = new((uint)bvhNode.AABB.Size.X, (uint)bvhNode.AABB.Size.Y),
                 LocalTransform = Transform2D.CreateTranslated(bvhNode.AABB.Position.X, -bvhNode.AABB.Position.Y)
             }
         };
