+++ source/Age/RenderPasses/CanvasRenderGraphPass.cs
@@ -151,7 +151,7 @@ public sealed class CanvasRenderGraphPass : CanvasBaseRenderGraphPass
         this.RenderPass.Dispose();
     }
 
-    protected override void ExecuteCommand(RenderPipelines resource, RectCommand command, in Size<float> viewport)
+    protected override void ExecuteCommand(RenderPipelines resource, RectCommand command, in Size<uint> viewport)
     {
         if (!this.UniformSets.TryGetValue(command.TextureMap.Texture, out var uniformSet))
         {
